target - SortMergeJoin detail - Extract the part that performs the join into the SortMergeJoinPerformer class reason - Extract the portion that performs join because the class is overweight due to too many methods
